backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
setuptools>=65.0.0
openpyxl==3.1.2
reportlab==4.0.7
django-redis==5.4.0

